+++ lizard_graph/views.py
@@ -1,10 +1,13 @@
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.txt.
+"""
+"""
 import datetime
 import iso8601
 import logging
 
 from django.views.generic.base import View
 from django.http import HttpResponse
+from django.utils import simplejson as json
 
 from lizard_fewsnorm.models import Series
 from lizard_fewsnorm.models import GeoLocationCache
@@ -18,14 +21,14 @@ from timeseries import timeseries
 logger = logging.getLogger(__name__)
 
 
-class LineGraph(NensGraph):
+class DateGridGraph(NensGraph):
     """
-    Standard line graph.
+    Standard graph with a grid and dates on the x-axis.
 
-    Inspired by lizard-map adapter.graph, but it is more generic."""
-
+    Inspired by lizard-map adapter.graph, but it is more generic.
+    """
     def __init__(self, **kwargs):
-        super(LineGraph, self).__init__(**kwargs)
+        super(DateGridGraph, self).__init__(**kwargs)
         self.axes = self.figure.add_subplot(111)
         self.axes.grid(True)
 
@@ -44,7 +47,7 @@ class LineGraph(NensGraph):
 
         if handles and labels:
             nitems = len(handles)
-            ncol = min(nitems, 3)
+            ncol = min(nitems, 2)
             # What comes next is an educated guess on the amount of
             # characters that can be used without collisions in the legend.
             ntrunc = int((self.width / ncol - 24) / 10)
@@ -60,23 +63,114 @@ class LineGraph(NensGraph):
                 fancybox=True,
                 shadow=True,)
 
-
-class GraphViewMixin(object):
-    """
-    """
-
-    def _time_series_from_request(self, request):
+    def line_from_single_ts(self, single_ts, color, graph_item):
         """
-        - location_id
-        - parameter_id
-        - module_id
+        Draw line(s) from a single timeseries.
+
+        Color is a matplotlib color, i.e. 'blue', 'black'
+
+        Graph_item can contain an attribute 'layout'.
+        """
+        title = '%s - %s (%s)' % (
+            single_ts.location_id, single_ts.parameter_id, single_ts.units)
+
+        dates, values, flag_dates, flag_values = single_ts.dates_values()
+
+        if not values:
+            return
+
+        # Line
+        self.axes.plot(
+            dates, values, "-", color=color, lw=2, label=title)
+
+        # Flags
+        self.axes.plot(
+            flag_dates, flag_values, "o-", color='red',
+            label=title + ' flags')
+
+    def bar_from_single_ts(self, bar_status, single_ts, color, graph_item):
+        """
+        Draw bars.
+
+        TODO: implement
+
+        Graph_item can contain an attribute 'layout'.
+        """
+
+        # Make seconds from fews timesteps.
+        TIME_STEPS = {'SETS1440': 1440 * 60}
+
+        dates, values, flag_dates, flag_values = single_ts.dates_values()
+
+        if not values:
+            return
+
+        self.axes.bar(dates, values, edgecolor='grey', width=60, label='test')
+
+
+class TimeSeriesViewMixin(object):
+    """
+    A mixin for a view that uses fewsnorm timeseries.
+    """
+
+    def _time_series_from_graph_item(
+        self, dt_start, dt_end, fews_norm_source_slug,
+        location_id=None, parameter_id=None, module_id=None, type=None):
+        """
+        - fews_norm_source_slug
+        - location_id (optional)
+        - parameter_id (optional)
+        - module_id (optional)
+        """
+
+        # fews_norm_source_slug = request.GET.get('fews_norm_source_slug', None)
+        # location_id = request.GET.get('location_id', None)  # '111.1'
+        # parameter_id = request.GET.get('parameter_id', None)  # 'ALMR110'
+        # module_id = request.GET.get('module_id', None)
+        # fews_norm_source = FewsNormSource.objects.get(
+        #     slug=fews_norm_source_slug)
+        fews_norm_source = FewsNormSource.objects.get(
+            slug=fews_norm_source_slug)
+
+        series = Series.objects.using(fews_norm_source.database_name).all()
+        if location_id is not None:
+            series = series.filter(location__id=location_id)
+        if parameter_id is not None:
+            series = series.filter(parameter__id=parameter_id)
+        if module_id is not None:
+            series = series.filter(module__id=module_id)
+
+        ts = timeseries.TimeSeries.as_dict(series, dt_start, dt_end)
+        return ts
+
+
+class GraphView(View, TimeSeriesViewMixin):
+    """
+    Draw standard line graph based on provided input.
+
+    Example request. Lizard-graph is mounted under 'graph', the source
+    slug is 'test':
+
+    http://127.0.0.1:8000/graph/?dt_start=2011-02-11%2000:00:00&dt_end=2011-11-11%2000:00:00&item={%22fews_norm_source_slug%22:%22test%22,%22location_id%22:%22111.1%22,%22parameter_id%22:%22ALMR110%22,%22type%22:%22line%22}
+    """
+    def get(self, request, *args, **kwargs):
+        """
+        Input:
         - dt_start
         - dt_end
-        """
-        location_id = request.GET.get('location_id', None)  # '111.1'
-        parameter_id = request.GET.get('parameter_id', None)  # 'ALMR110'
-        module_id = request.GET.get('module_id', None)
 
+        - item={fews_norm_source_slug:.., location_id:..,
+          parameter_id:.., module_id:.., type:.., arguments:..}
+        - type can be 'line', 'stacked-bar', 'vertical-line',
+          'horizontal-line', 'stacked-line' (see README).
+        - items are processed in order.
+        """
+        default_colors = ['green', 'blue', 'yellow', 'magenta', ]
+        graph = DateGridGraph()
+
+        graph_items_json = request.GET.getlist('item')
+        graph_items = [json.loads(graph_item_json)
+                       for graph_item_json in graph_items_json]
         start = request.GET.get('dt_start', None)
         end = request.GET.get('dt_end', None)
 
@@ -90,102 +184,23 @@ class GraphViewMixin(object):
         else:
             dt_end = iso8601.parse_date(end)
 
-        # if location_id is not None:
-        #     location_cache = GeoLocationCache.objects.filter(
-        #         ident=location_id)[0]
-        #     db_name = location_cache.fews_norm_source.database_name
-
-        series = Series.objects.using(self.fews_norm_source.database_name).all()
-        if location_id is not None:
-            series = series.filter(location__id=location_id)
-        if parameter_id is not None:
-            series = series.filter(parameter__id=parameter_id)
-        if module_id is not None:
-            series = series.filter(module__id=module_id)
-
-        ts = timeseries.TimeSeries.as_dict(series, dt_start, dt_end)
-        return ts
-
-    def _fews_norm_source_from_slug(self, slug):
-        return FewsNormSource.objects.get(slug=slug)
-
-
-class LineGraphView(View, GraphViewMixin):
-    """
-    Draw standard line graph based on provided input.
-
-    Example request. Lizard-graph is mounted under 'graph', the source
-    slug is 'test':
-
-    http://127.0.0.1:8000/graph/test/line/?location_id=111.1&parameter_id=ALMR110&dt_start=2010-11-11%2000:00:00&dt_end=2011-11-11%2000:00:00
-
-    """
-    def get(self, request, *args, **kwargs):
-        """
-        Input:
-
-        - fews_norm_source_slug (kwargs)
-        - location_id
-        - parameter_id
-        - module_id
-        - dt_start
-        - dt_end
-
-        TODO:
-        - implement all kinds of lines, max, min, etc.
-        """
-        self.fews_norm_source = self._fews_norm_source_from_slug(
-            kwargs['fews_norm_source_slug'])
-
-        colors = ['green', 'blue']
-        graph = LineGraph()
-
-        ts = self._time_series_from_request(request)
-
         color_index = 0
+        # bar_status is to keep track of the height of stacked bars.
+        bar_status = {}
+        for index, graph_item in enumerate(graph_items):
+            ts = self._time_series_from_graph_item(
+                dt_start, dt_end, **graph_item)
+            item_type = graph_item['type']
 
-        for (loc, par), single_ts in ts.items():
-            # print loc, par, single_ts
-            dates = []
-            values = []
-            flag_dates = []
-            flag_values = []
-            event_items = sorted(single_ts.events.items(),
-                                 key=lambda item: item[0])
-            for timestamp, (value, flag, comment) in event_items:
-                if value is not None:
-                    dates.append(timestamp)
-                    values.append(value)
+            for (loc, par), single_ts in ts.items():
+                color = default_colors[color_index]
+                if item_type == 'line':
+                    graph.line_from_single_ts(single_ts, color, graph_item)
+                elif item_type == 'bar':
+                    bar_status = graph.bar_from_single_ts(
+                        bar_status, single_ts, color, graph_item)
 
-                    # Flags:
-                    # 0: Original/Reliable
-                    # 1: Corrected/Reliable
-                    # 2: Completed/Reliable
-                    # 3: Original/Doubtful
-                    # 4: Corrected/Doubtful
-                    # 5: Completed/Doubtful
-                    # 6: Missing/Unreliable
-                    # 7: Corrected/Unreliable
-                    # 8: Completed/Unreliable
-                    # 9: Missing value
-                    if flag > 2:
-                        flag_dates.append(timestamp)
-                        flag_values.append(flag)
-
-            graph.axes.plot(
-                dates,
-                values,
-                "-",
-                color=colors[color_index],
-                lw=2,
-                label="lijntje")
-            graph.axes.plot(
-                flag_dates,
-                flag_values,
-                "o-",
-                color='red',
-                label="flag")
-            color_index = (color_index + 1) % len(colors)
+                color_index = (color_index + 1) % len(default_colors)
 
         graph.legend()
         return graph.png_response(